[server] 코드 정리
- 소스코드 헤더에 라이센스 표기
