+++ server/Ntreev.Crema.ServiceHosts.Http/Ntreev.Crema.ServiceHosts.Http.csproj
@@ -58,23 +58,13 @@
     <Reference Include="WindowsBase" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Commands\HttpCommandsService.cs" />
-    <Compile Include="Commands\HttpCommandsServiceHost.cs" />
-    <Compile Include="Commands\HttpCommandsServiceHostProvider.cs" />
-    <Compile Include="Commands\IHttpCommandsService.cs" />
-    <Compile Include="Commands\IHttpCommandsDataBaseService.cs" />
-    <Compile Include="Commands\IHttpCommandsUserService.cs" />
     <Compile Include="Controllers\Apis\V1\Commands\DataBaseApiController.cs" />
     <Compile Include="Controllers\Apis\V1\Commands\DefaultController.cs" />
     <Compile Include="Controllers\CremaApiController.cs" />
     <Compile Include="CremaAuthorizeAttribute.cs" />
-    <Compile Include="ErrorHandleBehavior.cs" />
+    <Compile Include="CremaExceptionHandler.cs" />
     <Compile Include="Extensions\ObjectExtensions.cs" />
     <Compile Include="HttpCommandServiceHost.cs" />
-    <Compile Include="HttpDescriptorService.cs" />
-    <Compile Include="HttpDescriptorServiceHost.cs" />
-    <Compile Include="HttpDescriptorServiceHostProvider.cs" />
-    <Compile Include="IHttpDescriptorService.cs" />
     <Compile Include="MefDependencyResolver.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Properties\Resources.Designer.cs">
@@ -92,8 +82,6 @@
     <Compile Include="Responses\Commands\ContainsDataBaseResponse.cs" />
     <Compile Include="Responses\Commands\IsDataBaseLoadedResponse.cs" />
     <Compile Include="Responses\Commands\LoginResponse.cs" />
-    <Compile Include="Responses\DescriptorService\IsOnlineResponse.cs" />
-    <Compile Include="Responses\DescriptorService\VersionResponse.cs" />
     <Compile Include="Responses\ErrorResponse.cs" />
     <Compile Include="Users\IHttpUserService.cs" />
     <Compile Include="Users\HttpUserService.cs" />
